linux/vboxsf: Another build fix (generic code now). bugref:9172
svn:sync-xref-src-repo-rev: r129583

+++ include/VBox/VBoxGuestLibSharedFoldersInline.h
@@ -1,4 +1,4 @@
-/* $Id: VBoxGuestLibSharedFoldersInline.h 77858 2019-03-24 00:10:24Z knut.osmundsen@oracle.com $ */
+/* $Id: VBoxGuestLibSharedFoldersInline.h 77881 2019-03-26 15:02:05Z knut.osmundsen@oracle.com $ */
 /** @file
  * VBoxGuestLib - Shared Folders Host Request Helpers (ring-0).
  */
@@ -1404,7 +1404,7 @@ DECLINLINE(int) VbglR0SfHostReqReadLinkContigSimple(SHFLROOT idRoot, const char 
     if (cchPath < _64K - 1)
     {
         VBOXSFREADLINKREQ *pReq = (VBOXSFREADLINKREQ *)VbglR0PhysHeapAlloc(RT_UOFFSETOF(VBOXSFREADLINKREQ, StrPath.String)
-                                                                           + SHFLSTRING_HEADER_SIZE + cchPath);
+                                                                           + SHFLSTRING_HEADER_SIZE + (uint32_t)cchPath);
         if (pReq)
         {
             pReq->StrPath.u16Length = (uint16_t)cchPath;
